Extract common logic for IntLong benchmarks
